+++ packages/map-template/src/components/Map/GoogleMapsMap/GoogleMapsMap.jsx
@@ -5,6 +5,8 @@ import { Loader as GoogleMapsApiLoader } from '@googlemaps/js-api-loader';
 import gmApiKeyState from '../../../atoms/gmApiKeyState';
 import primaryColorState from '../../../atoms/primaryColorState';
 import gmMapIdState from '../../../atoms/gmMapIdState';
+import bearingState from '../../../atoms/bearingState';
+import pitchState from '../../../atoms/pitchState';
 
 /**
  * Takes care of instantiating a MapsIndoors Google Maps MapView.
@@ -23,6 +25,8 @@ function GoogleMapsMap({ onMapView, onPositionControl }) {
     const [hasPositionControl, setHasPositionControl] = useState(false);
     const mapsIndoorsInstance = useRecoilValue(mapsIndoorsInstanceState);
     const primaryColor = useRecoilValue(primaryColorState);
+    const bearing = useRecoilValue(bearingState);
+    const pitch = useRecoilValue(pitchState);
 
     useEffect(() => {
         const loader = new GoogleMapsApiLoader({
@@ -41,8 +45,10 @@ function GoogleMapsMap({ onMapView, onPositionControl }) {
                 // Always set a center so the map so the bounds or center can be read from the start.
                 center: { lat: 0, lng: 0 },
                 // Set a large zoom so we prevent a "zoom 0 glitch" (showing the whole globe temporarily)
-                zoom: 21, 
-                mapId: gmMapId
+                zoom: 21,
+                mapId: gmMapId,
+                heading: !isNaN(parseInt(bearing)) ? parseInt(bearing) : 0,
+                tilt: !isNaN(parseInt(pitch)) ? parseInt(pitch) : 0,
             };
 
             const mapViewInstance = new window.mapsindoors.mapView.GoogleMapsView(mapViewOptions);
@@ -62,7 +68,6 @@ function GoogleMapsMap({ onMapView, onPositionControl }) {
         if (mapsIndoorsInstance && mapView && google && !hasPositionControl) {
             const myPositionButtonElement = document.createElement('mi-my-position');
             myPositionButtonElement.mapsindoors = mapsIndoorsInstance;
-
             mapView.getMap().controls[google.maps.ControlPosition.RIGHT_TOP].push(myPositionButtonElement);
             setHasPositionControl(true);
             onPositionControl(myPositionButtonElement);